requirements.txt
requests
opencv-python-headless
pillow
python-dotenv